% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020ctl}
\alias{pcod2020ctl}
\title{pcod.ctl iscam input file for pcod 2020 assessment}
\format{
\subsection{\code{pcod2020ctl}}{

A list with 8 items:
\describe{
\item{num.params}{Number of leading parameters in the model. Does not include catchability parameters (q)}
\item{params}{A matrix $num.params rows x 7 columns.

Parameters (rows):
\itemize{
\item \code{log_ro} log unfished recruitment.
\item \code{steepness} steepness.
\item \code{log_m} log natural mortality.
\item \code{log_avrec} log average recruitment.
\item \code{log_recinit} log average of initial recruitments to fill first year if population is not unfished at syr.
\item \code{rho} EIV: fraction of the total variance associated with observation error.
\item \code{kappa (varphi)} EIV: total precision (inverse of variance) of the total error.
}
Settings (columns):
\itemize{
\item \code{ival} starting value
\item \code{lb} lower bound (not sure if this can be used in tmb)
\item \code{ub} upper bound (not sure if this can be used in tmb)
\item \code{phz} phase of estimation (in ADMB). Won't need this in tmb but need to check how to fix a par
\item \code{prior} prior type:
\itemize{
\item 0 = uniform      (0,0),
\item 1 =  normal       (p1=mu,p2=sig),
\item 2 =  lognormal    (p1=log(mu),p2=sig),
\item 3 =  beta         (p1=alpha,p2=beta),
\item 4 =  gamma        (p1=alpha,p2=beta)
}
\item \code{p1} prior distribution parameters, see \code{prior}
\item \code{p2} prior distribution parameters, see \code{prior}}
}
\item{num.indices}{Number of indexes of abundance (surveys only). DEPRECATE THIS AND USE DAT$NIT}
\item{surv.q}{Controls for estimating catachability coefficients. Matrix 3 x num.indices (i.e., 1 column per index).
Settings (rows):
\itemize{
\item 1. Prior type:
\itemize{
\item  0 = uniform
\item  1 = normal on log q
\item  2 = random walk in q
}
\item 2. Prior mean
\item 3. Prior sd}
}
\item{fit.mean.weight}{Switch for fitting to annual mean weights. 0=FALSE, 1=TRUE.}
\item{num.mean.weight}{Number of mean weight series}
\item{weight.sig}{SD for mean weight likelihood}
\item{misc}{13 miscellaneous controls:
\itemize{
\item    1.  verbose ADMB output (0=off, 1=on)
\item    2.  recruitment model (1=beverton-holt, 2=ricker)
\item    3.  std in observed catches in first phase.
\item    4.  std in observed catches in last phase.
\item    5.  assume unfished equilibrium in first year (0=FALSE, 1=TRUE, 2 = At equilibrium with F in first year)
\item    6.  Maternal effects multiplier
\item    7.  Mean fishing mortality for regularizing the estimates of Ft
\item    8.  std in mean fishing mortality in first phase
\item    9.  std in mean fishing mortality in last phase
\item    10. phase for estimating m_deviations (use -1 to turn off mdevs)
\item    11. std in deviations for natural mortality
\item    12. number of estimated nodes for deviations in natural mortality
\item    13. fraction of total mortality that takes place prior to spawning}
}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020ctl
}
\description{
Control file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
