% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020pfc}
\alias{pcod2020pfc}
\title{pcod.pfc iscam input file for pcod 2020 assessment}
\format{
\subsection{\code{pcod2020pfc}}{

A list with 5 items:
\describe{
\item{num.tac}{}
\item{tac.vec}{}
\item{num.ctl.options}{}
\item{ctl.options}{}
\item{eof}{end of file check}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020pfc
}
\description{
projection file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
