% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020ctl}
\alias{pcod2020ctl}
\title{pcod.ctl iscam input file for pcod 2020 assessment}
\format{
\subsection{\code{pcod2020ctl}}{

A list with 12 items:
\describe{
\item{num.params}{number of parameters in the model}
\item{params}{matrix of parameters initial guesses, bounds, estimation phases and prior parameters}
\item{age.size}{matrix of age composition controls -add more description }
\item{sel}{selectivity control matrices, 10 rows and ncols=pcod2020dat$ngear}
\item{start.yr.time.block}{}
\item{num.indices}{number of indexes of abundance}
\item{surv.q}{controls for estimating catachability coefficients}
\item{fit.mean.weight}{}
\item{num.mean.weight.cv}{}
\item{weight.sig}{}
\item{misc}{miscellaneous controls}
\item{eof}{end of file check}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020ctl
}
\description{
Control file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
