% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020ctl}
\alias{pcod2020ctl}
\title{pcod.ctl iscam input file for pcod 2020 assessment}
\format{
\subsection{\code{pcod2020ctl}}{

A list with 12 items:
\describe{
\item{num.params}{Number of leading parameters in the model. Does not include catchability parameters (q)}
\item{params}{A matrix $num.params rows x 7 columns.
Parameters (rows): 1. log_ro = log unfished recruitment,
2. steepness = steepness,
3. log_m = log natural mortality,
4. log_avrec = log average recruitment,
5. log_recinit = log average of initial recruitments to fill first year,
6. rho = EIV: fraction of the total variance associated with observation error,
7. kappa (varphi) = EIV: total precision (inverse of variance) of the total error.
Settings (columns):
1. ival = starting value
2. lb = lower bound (not sure if this can be used in tmb)
3. ub = upper bound (not sure if this can be used in tmb)
4. phz = phase of estimation (in ADMB). Won't need this in tmb but need to check how to fix a par
5. prior = prior type:
0 uniform      (0,0)
1 normal       (p1=mu,p2=sig)
2 lognormal    (p1=log(mu),p2=sig)
3 beta         (p1=alpha,p2=beta)
4 gamma        (p1=alpha,p2=beta)
6. p1 = prior distribution parameters, see above
7. p2 = prior distribution parameters, see above}
\item{num.indices}{Number of indexes of abundance (surveys only). DEPRECATE THIS AND USE DAT$NIT}
\item{surv.q}{Controls for estimating catachability coefficients. Matrix 3 x num.indices (i.e., 1 column per index).
Settings (rows):
1. Prior type:
0 uniform
1 normal on log q
2 random walk in q
2. Prior mean
3. Prior sd}
\item{fit.mean.weight}{Switch for fitting to annual mean weights. 0=FALSE, 1=TRUE.}
\item{num.mean.weight}{Number of mean weight series}
\item{weight.sig}{SD for mean weight likelihood}
\item{misc}{13 miscellaneous controls:
1  -verbose ADMB output (0=off, 1=on)
2  -recruitment model (1=beverton-holt, 2=ricker)
3  -std in observed catches in first phase.
4  -std in observed catches in last phase.
5  -Assume unfished equilibrium in first year (0=FALSE, 1=TRUE, 2 = AT EQUILIBRIUM WITH FISHING MORTALITY IN SYR - IMPLEMENTED ONLY IN DELAY DIFF MODEL)
6  -Maternal effects multiplier
7  -Mean fishing mortality for regularizing the estimates of Ft
8  -std in mean fishing mortality in first phase
9  -std in mean fishing mortality in last phase
10 -phase for estimating m_deviations (use -1 to turn off mdevs)
11 -std in deviations for natural mortality
12 -number of estimated nodes for deviations in natural mortality
13 -fraction of total mortality that takes place prior to spawning}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020ctl
}
\description{
Control file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
