% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020dat}
\alias{pcod2020dat}
\title{pcod.dat iscam input file, filtered to only include inputs relevant to delay-difference model}
\format{
\subsection{\code{pcod2020dat}}{

A list with 25 items:
\describe{
\item{gearNames}{Names of the separate gears considered in the model, includes commercial and survey}
\item{syr}{Start year of data}
\item{nyr}{End year of data}
\item{sage}{Start age -- might not use this, just use kage}
\item{nage}{Maximum age}
\item{ngear}{Number of gears, includes commercial and survey}
\item{alloc}{Allocation of catch between gears, CHECK: Catch allocation: 0 < alloc <= 1 if commercial, 0 if survey}
\item{linf}{Growth, von Bertalanffy Linf}
\item{k}{Growth, von Bertalanffy k}
\item{to}{Growth, von Bertalanffy t0}
\item{lwscal}{Length-weight "a" (scale) parameter}
\item{lwpow}{Length-weight "b" (power) parameter}
\item{kage}{Knife-edge age-at-recruitment -- should be same as sage}
\item{alpha.g}{Growth, Ford-Walford alpha}
\item{rho.g}{Growth, Ford-Walford rho}
\item{wk}{Weight at age of recruitment}
\item{nctobs}{Number of catch observations}
\item{catch}{Catch data matrix, nctobs rows, 4 columns:}
\itemize{
\item year = Year
\item gear = Gear index from ngear
\item type:
\itemize{
\item 1 = catch in weight
\item 2 = catch in numbers
}
\item value (tonnes).
}
\item{nit}{Number of survey indexes}
\item{nitnobs}{Number of observations for each index of abundance. A vector of length nit.}
\item{survtype}{Survey type, a vector of length nit.
Type:
\itemize{
\item 1 = survey is proportional to numbers
\item 2 = survey is proportional to biomass
\item 3 = survey is proportional to spawning biomass (e.g., a spawn survey)}
}
\item{indices}{List of length nit of matrices nitnobs\link{i} rows x 5 columns:
\itemize{
\item \code{iyr} = Survey year, omits missing years
\item \code{it} = Survey value
\item \code{gear} = Gear index from ngear (includes both commercial and surveys)
\item \code{wt} = 1/CV (used as precision to multiplicatively weight observations)
\item \code{timing} = the fraction of total mortality that has occurred prior to survey. Usually zero.}
}
\item{nmeanwt}{Number of mean weight series}
\item{nmeanwtobs}{Number of observations for mean weight of the catch, a vector of length nmeanwt}
\item{meanwtdata}{Annual commercial mean weights (kg). A matrix with nmeanwtobs rows x 4 columns:
\itemize{
\item \code{year} = Year, omits missing years
\item \code{meanwt} = Mean weight value (kg)
\item \code{gear} = gear index from ngear
\item \code{wt} = fixed sd to use in likelihood. Usually 0.2 or 0.25, but subject to sensitivity analysis.
\item \code{timing} = the fraction of total mortality that has occurred prior to survey. Usually zero. Set same as for survey.}
}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020dat
}
\description{
Data file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
