% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.R
\docType{data}
\name{pcod2020dat}
\alias{pcod2020dat}
\title{pcod.dat iscam input file}
\format{
\subsection{\code{pcod2020dat}}{

A list with 44 items:
\describe{
\item{hasGearNames}{}
\item{gearNames}{Names of the separate gears considered in the model}
\item{hasAgeCompN}{Logical}
\item{narea}{number of areas in the assessment}
\item{ngroup}{number of groups in the assessment}
\item{nsex}{number of sex in the assessment}
\item{syr}{start year for the assessment model}
\item{nyr}{end year for the assessment model}
\item{sage}{start age of assessment}
\item{nage}{maximum age in the assessment}
\item{ngear}{number of gears in the assessment}
\item{alloc}{allocation of catch between gears}
\item{linf}{Von Bertallanffy Linf growth parameter estimate}
\item{k}{Von Bertallanffy k growth parameter estimate}
\item{to}{Von Bertallanffy t0 parameter estimate}
\item{lwscal}{length-weight "a" (scale) parameters}
\item{lwpow}{length-weight "b" (power) parameters}
\item{age50}{}
\item{sd50}{}
\item{usemat}{}
\item{matvec}{}
\item{dd.kage}{}
\item{dd.alpha.g}{}
\item{dd.rho.g}{}
\item{dd.wk}{}
\item{nctobs}{}
\item{catch}{catch data matrix, 65 rows and 7 columns}
\item{nit}{number of survey indexes}
\item{nitnobs}{number of observations for each index of abundance}
\item{survtype}{}
\item{indices}{list of length nit including matrices with index of abundance information}
\item{nagears}{number of gears with age data}
\item{nagearsvec}{}
\item{nagearssage}{start age in catch at age data}
\item{nagearsnage}{start age in catch at age data}
\item{eff}{effective sample size for age composition}
\item{agecompflag}{}
\item{agearsN}{}
\item{nwttab}{}
\item{nwtobs}{}
\item{nmeanwt}{}
\item{nmeanwtobs}{number of observations for mean weight of the catch}
\item{meanwtdata}{mean weight of the catch data}
\item{eof}{end of file check}
}
}
}
\source{
\url{https://github.com/pbs-assess/pacific-cod-2020}
}
\usage{
pcod2020dat
}
\description{
Data file used for input for iscam in the Pcod 2020 assessment
}
\keyword{datasets}
